service_orders/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.0
python-multipart==0.0.6
email-validator==2.1.0